% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{assign_course_date_by_date}
\alias{assign_course_date_by_date}
\title{Assign the date of the course that the survey response comes from.}
\usage{
assign_course_date_by_date(date, metadata_dates)
}
\arguments{
\item{date}{The date of the survey response.}
}
\value{
A character value.
}
\description{
Assign the date of the course that the survey response comes from.
}
\examples{
assign_course_date_by_date(
  c("2020-06-20", "2019-06-10", "2021-06-10"),
  get_course_dates("intro")
)
}
\keyword{internal}
