% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-teams.R
\name{assign_learners_to_groups}
\alias{assign_learners_to_groups}
\title{Assign learners into groups.}
\usage{
assign_learners_to_groups(data, group_names, score_cutoff = 3)
}
\arguments{
\item{data}{Survey data.}

\item{group_names}{Character vector that has the names to group into.}

\item{score_cutoff}{Point at which to split persons into "low" vs "high" skill (completely arbitrary).}
}
\value{
A data frame with team, name, and skill score.
}
\description{
Assign learners into groups.
}
