% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common-files.R
\name{copy_common_file}
\alias{copy_common_file}
\title{Copy commonly used files for r-cubed courses.}
\usage{
copy_common_file(file)
}
\arguments{
\item{template}{Name of common file.}
}
\value{
Path to created \verb{_variables.yml} file.
}
\description{
Copy commonly used files for r-cubed courses.
}
