% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre-course-misc.R
\name{copy_emails_for_slack_invite}
\alias{copy_emails_for_slack_invite}
\title{Paste emails from survey to send invites to join Slack.}
\usage{
copy_emails_for_slack_invite(emails)
}
\arguments{
\item{emails}{List of emails to format.}
}
\value{
Used for the side effect of pasting to the clipboard.
}
\description{
Paste emails from survey to send invites to join Slack.
}
