% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre-course-misc.R
\name{copy_instructors_to_groups_table}
\alias{copy_instructors_to_groups_table}
\title{Copy to clipboard the team-assigned instructors data as a Markdown table.}
\usage{
copy_instructors_to_groups_table(data, gh_org)
}
\arguments{
\item{data}{The data that contains the team names and the assigned
instructors.}

\item{gh_org}{The name of the course's GitHub organizaton, usually in the
form of \code{NAME-YYYY-MM}.}
}
\value{
Used for side effect of copying Markdown table to clipboard.
}
\description{
Copy to clipboard the team-assigned instructors data as a Markdown table.
}
