% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-github-org.R
\name{create_github_org}
\alias{create_github_org}
\title{Create a GItHub organization for a course based on a naming pattern of the course's ID.}
\usage{
create_github_org(id)
}
\arguments{
\item{id}{The identifier of the course based on \code{\link[=list_course_ids]{list_course_ids()}}.}
}
\value{
A character scalar with the name of the organization.
}
\description{
Create a GItHub organization for a course based on a naming pattern of the course's ID.
}
