% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-groups.R
\name{create_group_files}
\alias{create_group_files}
\title{Create the necessary files for group names for a workshop.}
\usage{
create_group_files(number_participants)
}
\arguments{
\item{number_participants}{Number of participants in the workshop.}
}
\value{
Invisibly returns TRUE. Side effect is to create a PDF file of each
group name to be printed for each table, an HTML file with name strips to
cut out for each participant for each day, as well as a text file with
the group names for use in other functions.
}
\description{
Create the necessary files for group names for a workshop.
}
