% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-teams.R
\name{create_group_names}
\alias{create_group_names}
\title{Create randomly generated group names.}
\usage{
create_group_names(max_name_length = 18)
}
\arguments{
\item{max_name_length}{Longest characters that the group name should be.}
}
\value{
A character vector of names.
}
\description{
Set a seed before hand to fix the names generated.
}
\examples{
create_group_names(20)
}
