% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project-setup.R
\name{create_inter_project_repo}
\alias{create_inter_project_repo}
\title{Create the team project for the intermediate workshop}
\usage{
create_inter_project_repo(path)
}
\arguments{
\item{path}{Path to where the project should be created. It will overwrite
any existing project.}
}
\value{
Invisibly returns the path to the created project.
}
\description{
Create the team project for the intermediate workshop
}
\examples{
\dontrun{
create_inter_project_repo("~/Desktop/dime-intermediate")
}
}
