% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre-course-misc.R
\name{create_team_project}
\alias{create_team_project}
\title{Clone a GitHub repo, setup the project with prodigenr, then commit and push.}
\usage{
create_team_project(repo_path, clone_directory = fs::path("~", "Desktop"))
}
\arguments{
\item{repo_path}{The GitHub style repo path (\code{orgname/reponame}).}

\item{clone_directory}{Where to clone the repository.}
}
\value{
Used for the side effects of creating the project, committing, and
pushing.
}
\description{
This function is used as a helper for \code{setup_team_repos()}.
}
