% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-md.R
\name{extract_code_block_with_label_string}
\alias{extract_code_block_with_label_string}
\title{Extract code blocks from a Markdown file that have a label that contains a specific string.}
\usage{
extract_code_block_with_label_string(md_file, string_match)
}
\arguments{
\item{md_file}{The markdown file to extract from.}

\item{string_match}{The string to look for that is inside the label. For
instance, if the label has word "function" inside the label.}
}
\description{
Extract code blocks from a Markdown file that have a label that contains a specific string.
}
