% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-feedback-types.R
\name{extract_feedback}
\alias{extract_feedback}
\alias{extract_feedback_quantitative}
\alias{extract_feedback_overall}
\alias{extract_feedback_sessions}
\title{Extract the quantitative responses from the feedback survey.}
\usage{
extract_feedback_quantitative(data)

extract_feedback_overall(data)

extract_feedback_sessions(data)
}
\arguments{
\item{data}{The feedback data.}
}
\value{
A tibble.
}
\description{
Extract the quantitative responses from the feedback survey.
}
\section{Functions}{
\itemize{
\item \code{extract_feedback_quantitative()}: Extract and tidy up the quantitative responses
from the feedback survey data.

\item \code{extract_feedback_overall()}: Extract and tidy up the overall comments
from the feedback survey data.

\item \code{extract_feedback_sessions()}: Extract and tidy up the session specific
comments from the feedback survey data.

}}
\examples{

\dontrun{
intro_feedback <- get_feedback_survey("intro")
general_feedback <- get_feedback_survey("general")

extract_feedback_quantitative(intro_feedback)
extract_feedback_quantitative(general_feedback)
}

}
