% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-participant-overview.R
\name{extract_participant_overview}
\alias{extract_participant_overview}
\title{Extract participant overview data from pre-course survey.}
\usage{
extract_participant_overview(data, id)
}
\arguments{
\item{data}{Pre-course survey data.}

\item{id}{The ID of the course, found by running \code{\link[=list_course_ids]{list_course_ids()}}.}
}
\value{
A \link[tibble:tibble]{tibble::tibble}.
}
\description{
Extract participant overview data from pre-course survey.
}
\examples{

\dontrun{
survey <- get_precourse_survey("intro")
extract_participant_overview(survey, "intro")
}
}
