% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch-precourse-survey.R
\name{extract_precourse}
\alias{extract_precourse}
\alias{extract_precourse_overview}
\alias{extract_precourse_feedback}
\title{Extract overview data from pre-course survey.}
\usage{
extract_precourse_overview(data, column_renaming_df)

extract_precourse_feedback(data, column_renaming_df)
}
\arguments{
\item{data}{Pre-survey data.}

\item{column_renaming_df}{Column that contains the renaming data.}
}
\value{
A tibble.
}
\description{
Extract overview data from pre-course survey.
}
\section{Functions}{
\itemize{
\item \code{extract_precourse_overview()}: Extract and tidy up the pre-course survey
participant overview data.

\item \code{extract_precourse_feedback()}: Extract and tidy up the pre-course feedback
data.

}}
