% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-precourse-feedback.R
\name{extract_preworkshop_feedback}
\alias{extract_preworkshop_feedback}
\title{Extract and tidy up the pre-workshop feedback data.}
\usage{
extract_preworkshop_feedback(data)
}
\arguments{
\item{data}{Pre-workshop survey data.}
}
\value{
A \link[tibble:tibble]{tibble::tibble}.
}
\description{
Extract and tidy up the pre-workshop feedback data.
}
\examples{
\dontrun{
survey <- get_preworkshop_survey("intro")
extract_preworkshop_feedback(survey)
}
}
