% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch-precourse-survey.R
\name{fetch_precourse}
\alias{fetch_precourse}
\alias{fetch_precourse_intro}
\alias{fetch_precourse_inter}
\alias{fetch_precourse_advanced}
\title{Fetch the pre-course survey responses.}
\usage{
fetch_precourse_intro(survey_id = Sys.getenv("INTRO_PRE_SURVEY_ID"))

fetch_precourse_inter(survey_id = Sys.getenv("INTERMEDIATE_PRE_SURVEY_ID"))

fetch_precourse_advanced(survey_id = Sys.getenv("ADVANCED_PRE_SURVEY_ID"))
}
\arguments{
\item{survey_id}{Google Forms ID for the survey.}
}
\value{
Data from survey, slightly tidied up.
}
\description{
Fetch the pre-course survey responses.
}
\section{Functions}{
\itemize{
\item \code{fetch_precourse_intro()}: Fetch the pre-course survey data for the \strong{introduction} course.

\item \code{fetch_precourse_inter()}: Fetch the pre-course survey data for the \strong{intermediate} course.

\item \code{fetch_precourse_advanced()}: Fetch the pre-course survey data for the \strong{advanced} course.

}}
\examples{
\dontrun{
fetch_precourse_intro()
fetch_precourse_advanced()
}

}
