% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch-precourse-survey.R
\name{fetch_survey_field_titles}
\alias{fetch_survey_field_titles}
\title{Fetch the column headers/titles and copy to the clipboard to tidy up.}
\usage{
fetch_survey_field_titles(survey_id)
}
\arguments{
\item{survey_id}{The ID for the Google Forms survey.}
}
\value{
Used for side effect of copying to clipboard.
}
\description{
Fetch the column headers/titles and copy to the clipboard to tidy up.
}
\examples{
\dontrun{
fetch_survey_field_titles(Sys.getenv("INTRO_PRE_SURVEY_ID"))
fetch_survey_field_titles(Sys.getenv("INTERMEDIATE_PRE_SURVEY_ID"))
fetch_survey_field_titles(Sys.getenv("ADVANCED_PRE_SURVEY_ID"))
fetch_survey_field_titles(Sys.getenv("INTRO_FEEDBACK_SURVEY_ID"))
fetch_survey_field_titles(Sys.getenv("INTERMEDIATE_FEEDBACK_SURVEY_ID"))
fetch_survey_field_titles(Sys.getenv("ADVANCED_FEEDBACK_SURVEY_ID"))
}
}
