% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/course-metadata.R
\name{get_course_metadata_field}
\alias{get_course_metadata_field}
\title{General purpose function}
\usage{
get_course_metadata_field(id, field)
}
\arguments{
\item{id}{The ID of the course, found by running \code{\link[=list_workshop_ids]{list_workshop_ids()}}.}

\item{field}{The "key" value (field) for the ID of the course, found by running \code{\link[=list_course_metadata_fields]{list_course_metadata_fields()}}.}
}
\value{
A (nested) list.
}
\description{
General purpose function
}
\examples{

get_course_metadata_field("intro", "events")
get_course_metadata_field("intro", "date")
}
