% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-feedback-survey.R
\name{get_feedback_survey}
\alias{get_feedback_survey}
\title{Get the responses from the feedback survey.}
\usage{
get_feedback_survey()
}
\value{
A \link[tibble:tibble]{tibble::tibble}.
}
\description{
Get the responses from the feedback survey.
}
\examples{
\dontrun{
get_feedback_survey()
}

}
