% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-precourse-survey.R
\name{get_preworkshop_survey}
\alias{get_preworkshop_survey}
\title{Get the (slightly cleaned) pre-workshop survey from Google Sheets.}
\usage{
get_preworkshop_survey(id)
}
\arguments{
\item{id}{The ID of the workshop, found by running \code{\link[=list_workshop_ids]{list_workshop_ids()}}.}
}
\value{
A \link[tibble:tibble]{tibble::tibble}.
}
\description{
Get the (slightly cleaned) pre-workshop survey from Google Sheets.
}
\examples{
\dontrun{
get_preworkshop_survey("intro")
}
}
