% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-schedule.R
\name{get_schedule_sessions}
\alias{get_schedule_sessions}
\title{Get the session schedule for the course.}
\usage{
get_schedule_sessions(id)
}
\arguments{
\item{id}{The ID of the course, found by running \code{\link[=list_course_ids]{list_course_ids()}}.}
}
\value{
A \link[tibble:tibble]{tibble::tibble}.
}
\description{
Get the session schedule for the course.
}
\examples{
get_schedule_sessions("intro")
get_schedule_sessions("inter")
}
