% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/course-metadata.R
\name{get_upcoming_course}
\alias{get_upcoming_course}
\title{Get the ID of the course that is upcoming.}
\usage{
get_upcoming_course()
}
\value{
A character scalar.
}
\description{
Get the ID of the course that is upcoming.
}
\examples{
get_upcoming_course()
}
