% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{get_upcoming_workshop}
\alias{get_upcoming_workshop}
\title{Get the ID of the workshop that is upcoming.}
\usage{
get_upcoming_workshop()
}
\value{
A character scalar.
}
\description{
Get the ID of the workshop that is upcoming.
}
\examples{
get_upcoming_workshop()
}
