% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{get_upcoming_workshop_dates}
\alias{get_upcoming_workshop_dates}
\title{Get the dates for the next workshop.}
\usage{
get_upcoming_workshop_dates(id)
}
\arguments{
\item{id}{The ID of the workshop, found by running \code{\link[=list_workshop_ids]{list_workshop_ids()}}.}
}
\value{
A character vector.
}
\description{
Get the dates for the next workshop.
}
\examples{

get_upcoming_workshop_dates("intro")
get_upcoming_workshop_dates("inter")
get_upcoming_workshop_dates("adv")
}
