% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{get_workshop_dates}
\alias{get_workshop_dates}
\title{Metadata on the dates the workshops have started on.}
\usage{
get_workshop_dates(id)
}
\arguments{
\item{id}{The ID of the workshop, found by running \code{\link[=list_workshop_ids]{list_workshop_ids()}}.}
}
\value{
Character vector of dates.
}
\description{
Metadata on the dates the workshops have started on.
}
\examples{
get_workshop_dates("inter")
get_workshop_dates("intro")
get_workshop_dates("adv")
}
