% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snippets.R
\name{install_snippets}
\alias{install_snippets}
\title{RStudio snippets to help with teaching}
\usage{
install_snippets()
}
\value{
Nothing. Used for the side effect of installing snippets for teaching.
}
\description{
RStudio snippets to help with teaching
}
