% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common-files.R
\name{list_common_files}
\alias{list_common_files}
\title{List files in the common directory.}
\usage{
list_common_files(regexp = NULL)
}
\arguments{
\item{regexp}{Regular expression for the file path.}
}
\value{
Vector of file paths
}
\description{
List files in the common directory.
}
\examples{
list_common_files("LICENSE.md")
}
