% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/course-metadata.R
\name{list_course_ids}
\alias{list_course_ids}
\title{List all the IDs for the courses.}
\usage{
list_course_ids()
}
\value{
A character vector.
}
\description{
List all the IDs for the courses.
}
\examples{
list_course_ids()
}
