% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/course-metadata.R
\name{list_course_metadata_fields}
\alias{list_course_metadata_fields}
\title{List all the metadata fields available in the first-level for each course.}
\usage{
list_course_metadata_fields()
}
\value{
A character vector.
}
\description{
List all the metadata fields available in the first-level for each course.
}
\examples{
list_course_metadata_fields()
}
