% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common-files.R
\name{list_template_files}
\alias{list_template_files}
\title{List files in the template directory.}
\usage{
list_template_files(regexp = NULL)
}
\arguments{
\item{regexp}{Regular expression for the file path.}
}
\value{
Vector of file paths.
}
\description{
List files in the template directory.
}
\examples{
list_template_files("planning-issue.md")
}
