% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{list_workshop_ids}
\alias{list_workshop_ids}
\title{List all the IDs for the workshops.}
\usage{
list_workshop_ids()
}
\value{
A character vector.
}
\description{
List all the IDs for the workshops.
}
\examples{
list_workshop_ids()
}
