% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{list_workshop_metadata_fields}
\alias{list_workshop_metadata_fields}
\title{List all the metadata fields available in the first-level for each workshop.}
\usage{
list_workshop_metadata_fields()
}
\value{
A character vector.
}
\description{
List all the metadata fields available in the first-level for each workshop.
}
\examples{
list_workshop_metadata_fields()
}
