% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{metadata_course_dates}
\alias{metadata_course_dates}
\title{Metadata on the dates the courses have started on.}
\usage{
metadata_course_dates(course_id = c("intro", "inter", "adv"))
}
\arguments{
\item{course_id}{ID for the course (found in the \code{courses.yaml} file).}
}
\value{
Character vector of dates.
}
\description{
Metadata on the dates the courses have started on.
}
\examples{
metadata_course_dates("inter")
metadata_course_dates("intro")
}
