% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_feedback}
\alias{plot_feedback}
\title{Plot the quantitative feedback from the survey data}
\usage{
plot_feedback(data)
}
\arguments{
\item{data}{The quantitative feedback data, as returned by \code{\link[=read_surveys]{read_surveys()}}.}
}
\value{
A \link[ggplot2:ggplot]{ggplot2::ggplot} object.
}
\description{
Plot the quantitative feedback from the survey data
}
