% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_overview}
\alias{plot_overview}
\title{Plot an overview of who the learners are.}
\usage{
plot_overview(data)
}
\arguments{
\item{data}{The pre-workshop survey data as returned by \code{\link[=read_surveys]{read_surveys()}}.}
}
\value{
A \link[ggplot2:ggplot]{ggplot2::ggplot} object.
}
\description{
Plot an overview of who the learners are.
}
