% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common-files.R
\name{read_common}
\alias{read_common}
\title{Read contents of a file in the common directory.}
\usage{
read_common(file)
}
\arguments{
\item{file}{Name of file in the common directory.}
}
\value{
Character vector.
}
\description{
Read contents of a file in the common directory.
}
\examples{
read_common("LICENSE.md")
}
