% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/course-metadata.R
\name{read_course_metadata}
\alias{read_course_metadata}
\title{Title}
\usage{
read_course_metadata()
}
\value{
A (heavily nested) list.
}
\description{
Title
}
\examples{
library(purrr)
read_course_metadata() |>
  purrr::map(names) |>
  unique()
}
