% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_surveys}
\alias{read_surveys}
\title{Read in specific survey results for all years for a workshop.}
\usage{
read_surveys(
  id,
  type = c("overview", "feedback-overall", "feedback-quantitative", "feedback-sessions")
)
}
\arguments{
\item{id}{The workshop ID, see \code{\link[=list_workshop_ids]{list_workshop_ids()}} for a list of them.}

\item{type}{The survey type, e.g. from the pre-workshop or post-session
feedback.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}}.
}
\description{
Read in specific survey results for all years for a workshop.
}
