% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common-files.R
\name{read_template_pkg_install}
\alias{read_template_pkg_install}
\title{Read the contents of the R package installation instructions template fiile.}
\usage{
read_template_pkg_install(course_type = c("intro", "inter", "advanced"))
}
\arguments{
\item{course_type}{The type of course.}
}
\value{
Character vector.
}
\description{
Read the contents of the R package installation instructions template fiile.
}
\examples{
read_template_pkg_install("intro")
}
