% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{read_workshop_metadata}
\alias{read_workshop_metadata}
\title{Read in the workshop's metadata (as a YAML) saved in the package's \verb{data/}
folder.}
\usage{
read_workshop_metadata()
}
\value{
A (heavily nested) list.
}
\description{
Read in the workshop's metadata (as a YAML) saved in the package's \verb{data/}
folder.
}
\examples{
read_workshop_metadata()
}
