% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{save_as_csv}
\alias{save_as_csv}
\title{Save data to CSV.}
\usage{
save_as_csv(data, path)
}
\arguments{
\item{data}{Survey data.}

\item{path}{Path where data should be saved to.}
}
\value{
Save a CSV file, returns tibble invisibly.
}
\description{
Save data to CSV.
}
