% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save-to-file.R
\name{save_to_csv}
\alias{save_to_csv}
\title{Save data to CSV.}
\usage{
save_to_csv(data, path)
}
\arguments{
\item{data}{Survey data.}

\item{path}{Path where data should be saved to.}
}
\value{
Save a CSV file, returns tibble invisibly.
}
\description{
Save data to CSV.
}
\examples{

library(dplyr)
general_feedback <- get_feedback_survey("general")
extract_feedback_quantitative(general_feedback)
extract_feedback_sessions(general_feedback) |>
  save_feedback_to_csv(c("course_id", "course_date", "date"))

}
