% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre-course-misc.R
\name{setup_team_repos}
\alias{setup_team_repos}
\title{Setup all team repositories to be ready for the assignment.}
\usage{
setup_team_repos(gh_org)
}
\arguments{
\item{gh_org}{The name of the course's GitHub organizaton, usually in the
form of \code{NAME-YYYY-MM}.}
}
\value{
Used for the side effect of selecting on all repos and setting up
projects for them.
}
\description{
Setup all team repositories to be ready for the assignment.
}
