% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%within\%}
\alias{\%within\%}
\title{Within date range, from lubridate}
\arguments{
\item{lhs}{A date.}

\item{rhs}{An interval object.}
}
\value{
A logical vector.
}
\description{
Within date range, from lubridate
}
\keyword{internal}
