% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin.R
\name{admin_create_planning_issue}
\alias{admin_create_planning_issue}
\title{Post a Course Planning issue on the course repo.}
\usage{
admin_create_planning_issue(id)
}
\arguments{
\item{id}{The ID of the course, found by running \code{\link[=list_workshop_ids]{list_workshop_ids()}}.}
}
\value{
A GitHub API response. Used for the side effects of posting an issue.
}
\description{
Post a Course Planning issue on the course repo.
}
