% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-github-org.R
\name{create_github_org}
\alias{create_github_org}
\title{Create a GitHub organization for a workshop.}
\usage{
create_github_org(id)
}
\arguments{
\item{id}{The identifier of the workshop based on \code{\link[=list_workshop_ids]{list_workshop_ids()}}.}
}
\value{
A character scalar with the name of the organization.
}
\description{
Create a GitHub organization for a workshop.
}
