% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-teams.R
\name{create_group_names}
\alias{create_group_names}
\title{Create randomly generated group names based on city coordinating from what3words.}
\usage{
create_group_names(number_groups)
}
\arguments{
\item{number_groups}{The number of group names to create.}
}
\value{
A character vector of names.
}
\description{
From all cities globally, get the \href{https://what3words.com/}{what3words}
coordinates to use as fun group names.
}
\examples{
create_group_names(20)
}
