% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-teams.R
\name{create_group_names}
\alias{create_group_names}
\title{Create randomly generated group names.}
\usage{
create_group_names(number_groups)
}
\arguments{
\item{number_groups}{The number of group names to create.}
}
\value{
A character vector of names.
}
\description{
Create randomly generated group names.
}
\examples{
\dontrun{
create_group_names(26)
}
}
