% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-teams.R
\name{create_teams}
\alias{create_teams}
\title{Create teams and assign learners to them.}
\usage{
create_teams(data, group_names, score_cutoff = 3)
}
\arguments{
\item{data}{Preworkshop survey data.}

\item{score_cutoff}{Point at which to split persons into "low" vs "high" skill (completely arbitrary).}

\item{team_names}{Character vector that has the names to group into.}
}
\value{
A \link[tibble:tibble]{tibble::tibble}.
}
\description{
Create teams and assign learners to them.
}
