% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-precourse-feedback.R
\name{extract_precourse_feedback}
\alias{extract_precourse_feedback}
\title{Extract and tidy up the pre-course feedback data.}
\usage{
extract_precourse_feedback(data)
}
\arguments{
\item{data}{Pre-course survey data.}
}
\value{
A \link[tibble:tibble]{tibble::tibble}.
}
\description{
Extract and tidy up the pre-course feedback data.
}
\examples{
\dontrun{
survey <- get_precourse_survey("intro")
extract_precourse_feedback(survey)
}
}
