% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/course-metadata.R
\name{get_course_dates}
\alias{get_course_dates}
\title{Metadata on the dates the courses have started on.}
\usage{
get_course_dates(id = list_course_ids())
}
\arguments{
\item{course_id}{ID for the course (found in the \code{courses.yaml} file).}
}
\value{
Character vector of dates.
}
\description{
Metadata on the dates the courses have started on.
}
\examples{
get_course_dates("inter")
get_course_dates("intro")
get_course_dates("adv")
}
