% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-feedback-survey.R
\name{get_feedback_survey}
\alias{get_feedback_survey}
\title{Get the responses from the feedback survey.}
\usage{
get_feedback_survey(id = "general")
}
\arguments{
\item{id}{The ID of the course, found by running \code{\link[=list_course_ids]{list_course_ids()}}.}
}
\value{
A \link[tibble:tibble]{tibble::tibble}.
}
\description{
Get the responses from the feedback survey.
}
\examples{

\dontrun{
get_feedback_survey("intro")
get_feedback_survey("general")
}

}
