% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/course-metadata.R
\name{get_upcoming_course_dates}
\alias{get_upcoming_course_dates}
\title{Title}
\usage{
get_upcoming_course_dates(id = list_course_ids())
}
\arguments{
\item{id}{}
}
\description{
Title
}
\examples{

get_upcoming_course_dates("intro")
get_upcoming_course_dates("inter")
get_upcoming_course_dates("adv")
}
