% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/course-metadata.R
\name{read_course_metadata}
\alias{read_course_metadata}
\title{Read in the course's metadata (as a YAML) saved in the package's \verb{data/}
folder.}
\usage{
read_course_metadata()
}
\value{
A (heavily nested) list.
}
\description{
Read in the course's metadata (as a YAML) saved in the package's \verb{data/}
folder.
}
\examples{
read_course_metadata()
}
